IGNITE-24955 wip
